Add files via upload
this is a project which i did when i was learning how to do codes and how to use MySQL database, ad this was my first project. we have to merge these files to diffrent folders according to their work and create a database using localhost and it is little tricky but is worth it.